--- fastcrypto-tbls/src/weight-reduction/src/util/knapsack/mod.rs
@@ -1,10 +1,0 @@
-mod dp;
-#[cfg(test)]
-mod dp_test;
-mod functions;
-#[cfg(test)]
-mod functions_test;
-
-pub use dp::DP;
-pub use functions::adversarial_tickets;
-pub use functions::is_valid;
