--- fastcrypto-tbls/src/weight-reduction/src/util/mod.rs
@@ -1,1 +1,0 @@
-pub mod knapsack;
